pyproject.toml
[tool.poetry]
name = "weathertelegrambot"
version = "0.1.0"
description = ""
authors = ["KarelinIvan <ivan.karelin.1993@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytelegrambotapi = "^4.26.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
